Add some simple speed control
Low Speed, Hi Speed, and turbo buton

+++ app/src/main/java/com/emerjbl/ultra8/Chip8Sound.kt
@@ -60,7 +60,7 @@ class Chip8Sound {
             track.reloadStaticData()
         }
         // The marker notification triggers the sound stop
-        val samples = (timer * SAMPLE_RATE / 60).toInt()
+        val samples = (timer * SAMPLE_RATE / 60)
         track.notificationMarkerPosition = samples
         track.play()
     }
